Moved fetching of stored Robots to the a presenter.
