+++ app/src/main/java/in/ivlabs/shayak/mainactivity/MainActivityPresenterInterface.kt
@@ -11,4 +11,11 @@ interface MainActivityPresenterInterface {
      * @return returns true if connection is established
      */
     fun connectToRobot(UUID : String) : Boolean
+
+    /**
+     *
+     *
+     * @return a list containing all saved robots
+     */
+    fun getRobotsList():List<MainActivityViewInterface.RobotViewData>
 }
