+++ app/src/main/java/in/ivlabs/shayak/view/MainActivity.kt
@@ -1,6 +1,5 @@
 package `in`.ivlabs.shayak.view
 
-import `in`.ivlabs.shayak.mainactivity.MainActivityPresenterInterface
 import `in`.ivlabs.shayak.mainactivity.MainActivityViewInterface
 import android.graphics.Color
 import android.os.Bundle
@@ -9,29 +8,29 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.ivlabs.shayak.R
 import com.mikepenz.fastadapter.FastAdapter
 import com.mikepenz.fastadapter.adapters.ItemAdapter
 import com.mikepenz.fastadapter.items.AbstractItem
+import kotlinx.android.synthetic.main.activity_main.*
 
 val Boolean.int
     get() = if (this) 1 else 0
 
-class MainActivity : AppCompatActivity(), MainActivityViewInterface,
-    MainActivityPresenterInterface {
+class MainActivity : AppCompatActivity(), MainActivityViewInterface {
+    val presenter = MainActivityPresenter(this)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        displayRobotList(getDummyRobotList())
+        val robotList = presenter.getRobotsList()
+        displayRobotList(robotList)
     }
 
     override fun displayRobotList(list: List<MainActivityViewInterface.RobotViewData>) {
         val robotAdapter = ItemAdapter<RobotItem>()
         val fastAdapter = FastAdapter.with(robotAdapter)
-        findViewById<RecyclerView>(R.id.main_recycler).adapter = fastAdapter
-        findViewById<RecyclerView>(R.id.main_recycler).layoutManager =
-            LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false)
+        main_recycler.adapter = fastAdapter
+        main_recycler.layoutManager = LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false)
         robotAdapter.add(
             list.sortedWith(Comparator { o1: MainActivityViewInterface.RobotViewData, o2: MainActivityViewInterface.RobotViewData -> o2.isFavorite.int - o1.isFavorite.int })
                 .map {
@@ -44,22 +43,6 @@ class MainActivity : AppCompatActivity(), MainActivityViewInterface,
         TODO("Not yet implemented")
     }
 
-    override fun connectToRobot(UUID: String): Boolean {
-        TODO("Not yet implemented")
-    }
-
-    /*Gets Dummy List of Robots to test RecyclerView Population*/
-    fun getDummyRobotList(): List<MainActivityViewInterface.RobotViewData> {
-        val robotList = List<MainActivityViewInterface.RobotViewData>(5) {
-            MainActivityViewInterface.RobotViewData(
-                "$it",
-                "Robot $it",
-                "Short Description of Robot $it",
-                it % 2 == 0
-            )
-        }
-        return robotList
-    }
 }
 
 class RobotItem(robot: MainActivityViewInterface.RobotViewData) :
@@ -70,7 +53,7 @@ class RobotItem(robot: MainActivityViewInterface.RobotViewData) :
     val uuid = robot.UUID
 
     class ViewHolder(view: View) : FastAdapter.ViewHolder<RobotItem>(view) {
-        var nameTextview = view.findViewById<TextView>(R.id.robot_card_title)
+        val nameTextview = view.findViewById<TextView>(R.id.robot_card_title)
         var descriptionTextview = view.findViewById<TextView>(R.id.robot_card_description)
         var favouriteImageview = view.findViewById<ImageView>(R.id.robot_card_favourite)
         override fun bindView(item: RobotItem, payloads: List<Any>) {
